requirements.txt
# Web framework
streamlit==1.29.0

# Azure OpenAI and related
openai==1.6.1
azure-identity==1.15.0

# LangChain components
langchain-openai==0.0.5
langgraph==0.0.19

# Elasticsearch
elasticsearch==8.11.1

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2

# RAG evaluation
ragas==0.0.18
datasets==2.16.1

# Environment and configuration
python-dotenv==1.0.0

# Logging and utilities
pathlib2==2.3.7
requests==2.31.0

# Additional dependencies
typing-extensions==4.9.0